reorganize builder
