+++ compiler/src/core/builder/single-builder.ts
@@ -1,7 +1,7 @@
 import type { Slice } from "@/core/layout";
-import { Block } from "../block";
-import { Direction } from "../direction";
 import { Builder } from "./builder";
+import { Block } from "./utils/block";
+import { Direction } from "./utils/direction";
 
 export class SingleBuilder extends Builder<"single"> {
 	protected override buildSlice(slice: Slice<"single">) {
