+++ compiler/src/core/builder/utils/block-placer.ts
