+++ compiler/src/core/builder/utils/block.ts
