+++ compiler/src/core/builder/utils/buffer.ts
