+++ compiler/src/core/builder/utils/cursor.ts
