+++ compiler/src/core/builder/utils/size.ts
